+++ tests/base.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
 from unittest import TestCase
+from doctest import Example
+from lxml.doctestcompare import LXMLOutputChecker
 
 from httmock import HTTMock, all_requests
 
@@ -71,6 +73,24 @@ class PlivoResourceTestCase(TestCase):
 
     def get_url(self, *args, **kwargs):
         return 'https://api.plivo.com/v1/Account/' + \
+               self.client.session.auth[0] + \
+               '/' + '/'.join([quote_plus(arg)
+                               for arg in args]) + '/?' + urlencode(kwargs)
+
+    def get_voice_url(self, *args, **kwargs):
+        return 'https://voice.plivo.com/v1/Account/' + \
+               self.client.session.auth[0] + \
+               '/' + '/'.join([quote_plus(arg)
+                               for arg in args]) + '/?' + urlencode(kwargs)
+
+    def get_voice_fallback1_url(self, *args, **kwargs):
+        return 'https://voice-usw1.plivo.com/v1/Account/' + \
+               self.client.session.auth[0] + \
+               '/' + '/'.join([quote_plus(arg)
+                               for arg in args]) + '/?' + urlencode(kwargs)
+
+    def get_voice_fallback2_url(self, *args, **kwargs):
+        return 'https://voice-use1.plivo.com/v1/Account/' + \
                self.client.session.auth[0] + \
                '/' + '/'.join([quote_plus(arg)
                                for arg in args]) + '/?' + urlencode(kwargs)
@@ -154,3 +174,12 @@ class PlivoPhlosResourceTestCase(TestCase):
                self.client.session.auth[0] + \
                '/' + '/'.join([quote_plus(arg)
                                for arg in args]) + '/?' + urlencode(kwargs)
+
+
+class PlivoXmlTestCase:
+
+    def assertXmlEqual(self, got, want):
+        checker = LXMLOutputChecker()
+        if not checker.check_output(want, got, 0):
+            message = checker.output_difference(Example("", want), got, 0)
+            raise AssertionError(message)