+++ tests/xml/test_sElement.py
@@ -1,9 +1,10 @@
 from unittest import TestCase
 
 from plivo import plivoxml
+from tests import PlivoXmlTestCase
 
 
-class SElementTest(TestCase):
+class SElementTest(TestCase, PlivoXmlTestCase):
     def test_set_methods(self):
         expected_response = '<Response><Speak><s><break strength="strong"/>' \
                             '<emphasis level="strong">This is Test</emphasis><lang xml:lang="it">This is ' \
@@ -76,4 +77,4 @@ class SElementTest(TestCase):
                 )
             )
         ).to_string(False)
-        self.assertEqual(response, expected_response)
+        self.assertXmlEqual(response, expected_response)